REFACTOR: Move adapter and persistence classes to infrastructure
Reorganized the project structure by moving adapter and persistence-related classes to the `infrastructure` package for better modularity and alignment. Updated import paths across the project to reflect the changes and maintain consistency. This improves maintainability and separation of concerns.
