+++ src/main/java/br/com/donor/donorsapi/adapters/mapper/DonorOverWeight.java
@@ -1,0 +1,12 @@
+package br.com.donor.donorsapi.adapters.mapper;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class DonorOverWeight {
+    private String gender;
+    private int total;
+    private double percent;
+}
