--- src/main/java/br/com/donor/donorsapi/adapters/model/DonorOverWeight.java
@@ -1,12 +1,0 @@
-package br.com.donor.donorsapi.adapters.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class DonorOverWeight {
-    private String gender;
-    private int total;
-    private double percent;
-}
