--- src/main/java/br/com/donor/donorsapi/adapters/persistence/entity/AgeImcData.java
@@ -1,13 +1,0 @@
-package br.com.donor.donorsapi.adapters.persistence.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class AgeImcData {
-    String ageRange;
-    double averageImc;
-}
