--- src/main/java/br/com/donor/donorsapi/adapters/persistence/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package br.com.donor.donorsapi.adapters.persistence.repository;
-
-import br.com.donor.donorsapi.adapters.persistence.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    Optional<UserEntity> findByUsername(String username);
-
-    Optional<UserEntity> findByEmail(String email);
-}
