--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/bmi/DonorAdapterBmiService.java
@@ -1,10 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.bmi;
-
-import br.com.donor.donorsapi.adapters.persistence.entity.AgeImcData;
-
-import java.util.List;
-
-public interface DonorAdapterBmiService {
-
-    List<AgeImcData> findAllAgeAndImc();
-}
