--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/overweight/DonorAdapterOverWeightService.java
@@ -1,10 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.overweight;
-
-import br.com.donor.donorsapi.adapters.model.DonorOverWeight;
-
-import java.util.List;
-
-public interface DonorAdapterOverWeightService {
-
-    List<DonorOverWeight> getOverWeight();
-}
