--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/overweight/DonorAdapterOverWeightServiceImpl.java
@@ -1,23 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.overweight;
-
-import br.com.donor.donorsapi.adapters.model.DonorOverWeight;
-import br.com.donor.donorsapi.domain.repository.DonorRepository;
-import br.com.donor.donorsapi.domain.service.donor.overweight.DonorOverWeightService;
-import br.com.donor.donorsapi.domain.service.donor.overweight.DonorOverWeightServiceImpl;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class DonorAdapterOverWeightServiceImpl implements DonorAdapterOverWeightService {
-    private final DonorOverWeightService service;
-
-    public DonorAdapterOverWeightServiceImpl(DonorRepository repository) {
-        this.service = new DonorOverWeightServiceImpl(repository);
-    }
-
-    @Override
-    public List<DonorOverWeight> getOverWeight() {
-        return service.getOverWeights();
-    }
-}
