--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/recipient/DonorAdapterRecipientService.java
@@ -1,9 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.recipient;
-
-import br.com.donor.donorsapi.domain.model.BloodTypeRecipientCount;
-
-import java.util.List;
-
-public interface DonorAdapterRecipientService {
-    List<BloodTypeRecipientCount> getDonorBloodType();
-}
