--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/state/DonorAdpterSericeState.java
@@ -1,10 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.state;
-
-import br.com.donor.donorsapi.domain.model.Donor;
-
-import java.util.List;
-
-public interface DonorAdpterSericeState {
-
-    List<Donor> getByState(String state);
-}
