--- src/main/java/br/com/donor/donorsapi/adapters/service/donor/yearaverage/DonorYearAverageService.java
@@ -1,10 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.donor.yearaverage;
-
-import br.com.donor.donorsapi.domain.model.DonorYearAverage;
-
-import java.util.List;
-
-public interface DonorYearAverageService {
-
-    List<DonorYearAverage> getYearAverage();
-}
