--- src/main/java/br/com/donor/donorsapi/adapters/service/user/UserUseCase.java
@@ -1,6 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.user;
-
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-public interface UserUseCase extends UserDetailsService {
-}
