--- src/main/java/br/com/donor/donorsapi/adapters/service/user/UserUseCaseImpl.java
@@ -1,20 +1,0 @@
-package br.com.donor.donorsapi.adapters.service.user;
-
-import br.com.donor.donorsapi.adapters.persistence.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserUseCaseImpl implements UserUseCase {
-    private final UserRepository repository;
-
-    public UserUseCaseImpl(UserRepository repository) {
-        this.repository = repository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return repository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("UserEntity not found"));
-    }
-}
