+++ src/main/java/br/com/donor/donorsapi/domain/service/donor/bmi/DonorImcService.java
@@ -1,6 +1,6 @@
 package br.com.donor.donorsapi.domain.service.donor.bmi;
 
-import br.com.donor.donorsapi.adapters.persistence.entity.AgeImcData;
+import br.com.donor.donorsapi.infrastructure.model.AgeImcData;
 
 import java.util.List;
