+++ src/main/java/br/com/donor/donorsapi/domain/service/donor/bmi/DonorImcServiceImpl.java
@@ -1,8 +1,8 @@
 package br.com.donor.donorsapi.domain.service.donor.bmi;
 
-import br.com.donor.donorsapi.adapters.persistence.DonorImcMapper;
-import br.com.donor.donorsapi.adapters.persistence.entity.AgeImcData;
+import br.com.donor.donorsapi.adapters.mapper.DonorImcMapper;
 import br.com.donor.donorsapi.domain.repository.DonorRepository;
+import br.com.donor.donorsapi.infrastructure.model.AgeImcData;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
