+++ src/main/java/br/com/donor/donorsapi/domain/service/donor/overweight/DonorOverWeightService.java
@@ -1,6 +1,6 @@
 package br.com.donor.donorsapi.domain.service.donor.overweight;
 
-import br.com.donor.donorsapi.adapters.model.DonorOverWeight;
+import br.com.donor.donorsapi.adapters.mapper.DonorOverWeight;
 
 import java.util.List;
