+++ src/main/java/br/com/donor/donorsapi/infrastructure/persistence/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package br.com.donor.donorsapi.infrastructure.persistence.repository;
+
+import br.com.donor.donorsapi.infrastructure.persistence.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByUsername(String username);
+
+}
