+++ src/main/java/br/com/donor/donorsapi/infrastructure/service/donor/bmi/DonorAdapterBmiService.java
@@ -1,0 +1,10 @@
+package br.com.donor.donorsapi.infrastructure.service.donor.bmi;
+
+import br.com.donor.donorsapi.infrastructure.model.AgeImcData;
+
+import java.util.List;
+
+public interface DonorAdapterBmiService {
+
+    List<AgeImcData> findAllAgeAndImc();
+}
