+++ src/main/java/br/com/donor/donorsapi/infrastructure/service/donor/overweight/DonorAdapterOverWeightService.java
@@ -1,0 +1,10 @@
+package br.com.donor.donorsapi.infrastructure.service.donor.overweight;
+
+import br.com.donor.donorsapi.adapters.mapper.DonorOverWeight;
+
+import java.util.List;
+
+public interface DonorAdapterOverWeightService {
+
+    List<DonorOverWeight> getOverWeight();
+}
