+++ src/main/java/br/com/donor/donorsapi/infrastructure/service/donor/recipient/DonorAdapterRecipientService.java
@@ -1,0 +1,9 @@
+package br.com.donor.donorsapi.infrastructure.service.donor.recipient;
+
+import br.com.donor.donorsapi.domain.model.BloodTypeRecipientCount;
+
+import java.util.List;
+
+public interface DonorAdapterRecipientService {
+    List<BloodTypeRecipientCount> getDonorBloodType();
+}
