+++ src/main/java/br/com/donor/donorsapi/infrastructure/service/user/UserUseCase.java
@@ -1,0 +1,6 @@
+package br.com.donor.donorsapi.infrastructure.service.user;
+
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+public interface UserUseCase extends UserDetailsService {
+}
