+++ src/main/java/br/com/donor/donorsapi/infrastructure/service/user/UserUseCaseImpl.java
@@ -1,0 +1,20 @@
+package br.com.donor.donorsapi.infrastructure.service.user;
+
+import br.com.donor.donorsapi.infrastructure.persistence.repository.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserUseCaseImpl implements UserUseCase {
+    private final UserRepository repository;
+
+    public UserUseCaseImpl(UserRepository repository) {
+        this.repository = repository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return repository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("UserEntity not found"));
+    }
+}
