+++ src/main/java/br/com/donor/donorsapi/web/controller/auth/dto/UserDto.java
@@ -1,6 +1,6 @@
 package br.com.donor.donorsapi.web.controller.auth.dto;
 
-import br.com.donor.donorsapi.adapters.persistence.entity.UserEntity;
+import br.com.donor.donorsapi.infrastructure.persistence.entity.UserEntity;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
